+++ src/main/java/com/biancodavide3/budgeting/db/repositories/BudgetRepository.java
@@ -15,7 +15,7 @@ import java.util.Optional;
 
 /*
 CREATE TABLE budgets (
-                         id SERIAL PRIMARY KEY,
+                         id BIGSERIAL PRIMARY KEY,
                          user_id INTEGER REFERENCES users(id) ON DELETE CASCADE,
                          month VARCHAR(7) NOT NULL, -- Format: YYYY-MM
                          total_budget NUMERIC(10, 2) NOT NULL,
@@ -24,12 +24,12 @@ CREATE TABLE budgets (
  */
 
 @Repository
-public interface BudgetRepository extends JpaRepository<BudgetEntity, Integer>, JpaSpecificationExecutor<BudgetEntity> {
-    List<BudgetEntity> findAllByUser_Id(Integer userId);
-    Page<BudgetEntity> findAllByUser_Id(Integer userId, Pageable pageable);
-    Optional<BudgetEntity> findByUser_IdAndMonth(Integer userId, YearMonth month);
-    List<BudgetEntity> findAllByUser_IdAndTotalBudgetGreaterThan(Integer userId, BigDecimal totalBudget);
-    Page<BudgetEntity> findAllByUser_IdAndTotalBudgetGreaterThan(Integer userId, BigDecimal totalBudget, Pageable pageable);
-    List<BudgetEntity> findAllByUser_IdAndTotalBudgetLessThan(Integer userId, BigDecimal totalBudget);
-    Page<BudgetEntity> findAllByUser_IdAndTotalBudgetLessThan(Integer userId, BigDecimal totalBudget, Pageable pageable);
+public interface BudgetRepository extends JpaRepository<BudgetEntity, Long>, JpaSpecificationExecutor<BudgetEntity> {
+    List<BudgetEntity> findAllByUser_Id(Long userId);
+    Page<BudgetEntity> findAllByUser_Id(Long userId, Pageable pageable);
+    Optional<BudgetEntity> findByUser_IdAndMonth(Long userId, YearMonth month);
+    List<BudgetEntity> findAllByUser_IdAndTotalBudgetGreaterThan(Long userId, BigDecimal totalBudget);
+    Page<BudgetEntity> findAllByUser_IdAndTotalBudgetGreaterThan(Long userId, BigDecimal totalBudget, Pageable pageable);
+    List<BudgetEntity> findAllByUser_IdAndTotalBudgetLessThan(Long userId, BigDecimal totalBudget);
+    Page<BudgetEntity> findAllByUser_IdAndTotalBudgetLessThan(Long userId, BigDecimal totalBudget, Pageable pageable);
 }
